requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pypdf>=4.0.0
faiss-cpu>=1.8.0
langchain>=0.2.0
langchain-community>=0.2.0
google-generativeai>=0.8.0
python-docx>=1.1.0
python-dotenv>=1.0.0
